backend/pyproject.toml
[project]
name = "content-management-backend"
version = "0.1.0"
description = "Educational Content Analysis System Backend"
requires-python = ">=3.11"
dependencies = [
    "django>=4.2.0",
    "psycopg2-binary>=2.9.0",
    "nanodjango>=0.8.0",
    "django-ninja>=1.0.0",
    "pyjwt>=2.8.0",
    "openai>=1.0.0",
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.2.0",
    "redis>=5.0.0",
    "rq>=1.15.0",
    "django-rq>=2.8.0",
    "python-dotenv>=1.0.0",
    "django-redis>=5.4.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-django>=4.5.2",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "pytest-asyncio>=0.21.0",
    "factory-boy>=3.3.0",
    "faker>=19.6.0",
    "freezegun>=1.2.2",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["tests"]
addopts = [
    "--reuse-db",
    "--nomigrations",
    "--cov=app",
    "--cov-report=html",
    "--cov-report=term-missing",
    "-v",
]
